Combat almost gets to menu

+++ src/main/java/legend/game/Scus94491BpeSegment_800e.java
@@ -1278,7 +1278,7 @@ public final class Scus94491BpeSegment_800e {
     GsInitCoordinate2(null, bigStruct.coord2_14);
     FUN_80021ca0(bigStruct.ObjTable_0c, bigStruct.tmd_8c.deref(), bigStruct.coord2_14, bigStruct.count_c8.get(), (short)(bigStruct.tmdNobj_ca.get() + 0x1L));
 
-    bigStruct.us_a0.set(0);
+    bigStruct.us_a0.set((short)0);
     bigStruct.ub_a2.set(0);
     bigStruct.ub_a3.set(0);
     bigStruct.ui_f4.set(0);
